api Change: methods can optionally return unsigned tx
For any function that required a keypair, it is possible to pass a
public key instead. In that case, the transaction that accomplishes the
requested function is returned instead of signed & sent. The point of
this is to handle cases where a transaction needs multiple signatures
before being broadcasted.

_Note: This was the behavior of the first oc-multisig releases, but was
disabled at some point to simplify the API. However, this is still
required in some cases._